move common implementation to AbstractBaseClient class
